add some error checking, fix bug when outputting function names that are MSIL tokens

+++ netil/tests/cc386/AN386.c
@@ -36,7 +36,7 @@ mailto::camille@bluegrass.net
 #include "rtti.h"
 
 /* pc-relative expressions not optimized */
-extern int nextlabel;
+extern long nextlabel;
 extern AMODE push[], pop[];
 extern OCODE *peep_tail;
 extern SYM *currentfunc;
